+++ src/components/common/Cart/styled.js
@@ -24,16 +24,6 @@ export const CartList = styled.ul`
 export const TitleCheck = styled.h4`
     text-align: center;
 `;
-export const ItemList = styled.li`
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    margin: 10px 0;
-`;
-export const CartImg = styled.img`
-    height: 60px;
-    width: 70px;
-`;
 export const SumCount = styled.div`
     display: flex;
     justify-content: flex-end;
