+++ src/components/common/CartProductItem/cartProductItem.jsx
@@ -1,0 +1,16 @@
+import React from 'react'
+import { inject, observer } from 'mobx-react'
+
+import {
+  CartImg, ItemList,
+} from './styled'
+
+export const CartProductItem = inject('cartStore')(observer(({ cartStore }) => (
+  <>
+    <ItemList>
+      <CartImg src='./cart/activiafin.png' alt='product item' />
+      <span>{cartStore.text}</span>
+      <span>$5</span>
+    </ItemList>
+  </>
+)))
