+++ src/components/common/CartProductItem/styled.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components'
+
+export const ItemList = styled.li`
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    margin: 10px 0;
+`;
+export const CartImg = styled.img`
+    height: 60px;
+    width: 70px;
+`;
