+++ src/components/common/index.js
@@ -2,7 +2,7 @@ import { Header } from './Header/header'
 import { Layout } from './Layout/layout'
 import { SEO } from './Seo/seo'
 import { CustomModal } from './Modal/modal'
-import { Cart } from './Cart/Cart'
+import { Cart } from './Cart/cart'
 import { Check } from './Check/check'
 
 export {
