+++ src/pages/index.js
@@ -2,6 +2,7 @@ import React from 'react'
 import { Link } from 'gatsby'
 
 import { Provider } from 'mobx-react'
+
 import { Layout } from '../components/common/Layout/layout'
 import { SEO } from '../components/common/Seo/seo'
 import { ProductList } from '../components/Main'
@@ -22,4 +23,4 @@ const IndexPage = () => (
 
 export default IndexPage
 
-stores.text = 'qwerty'
+stores.cartStore.text = 'asasassdasfasa'
